requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
fastapi==0.116.1
filelock==3.13.1
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyMatting==1.1.14
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
rembg==2.0.67
requests==2.32.4
rpds-py==0.26.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
setuptools==78.1.1
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
tifffile==2025.6.11
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
v==1
Werkzeug==3.1.3
wheel==0.45.1
